Move create_tables to DButils.
