+++ unit_tests/test_DButils.py
@@ -29,6 +29,19 @@ def remove_tmpfile(fname):
     os.remove(fname)
 
 
+class TestCreateTables(unittest.TestCase):
+    """Create the tables in a new DB"""
+
+    def test_create_tables(self):
+        td = tempfile.mkdtemp()
+        try:
+            working = os.path.join(td, 'emptyDB.sqlite')
+            DButils.create_tables(working)
+            dbu = DButils.DButils(working)
+            self.assertTrue(hasattr(dbu, 'File'))
+        finally:
+            shutil.rmtree(td)
+
 class TestSetupNoOpen(unittest.TestCase, dbp_testing.AddtoDBMixin):
     """
     master class for the setup and teardown, without opening db
